break: rename createUpdater to combine
