+++ packages/dataparcels/combine.js
@@ -1,0 +1,2 @@
+/* eslint-disable */
+module.exports = require('./lib/parcelData/combine.js').default;
