--- packages/dataparcels/createUpdater.js
@@ -1,2 +1,0 @@
-/* eslint-disable */
-module.exports = require('./lib/parcelData/createUpdater.js').default;
