+++ packages/react-dataparcels/__test__/Exports-test.js
@@ -4,7 +4,7 @@
 import Parcel from '../src/index';
 import Action from '../Action';
 import ChangeRequest from '../ChangeRequest';
-import createUpdater from '../createUpdater';
+import combine from '../combine';
 import deleted from '../deleted';
 import ParcelNode from '../ParcelNode';
 import arrange from '../arrange';
@@ -28,7 +28,7 @@ import asyncValue from '../asyncValue';
 import InternalParcel from 'dataparcels';
 import InternalAction from 'dataparcels/Action';
 import InternalChangeRequest from 'dataparcels/ChangeRequest';
-import InternalCreateUpdater from 'dataparcels/createUpdater';
+import InternalCreateUpdater from 'dataparcels/combine';
 import Internaldeleted from 'dataparcels/deleted';
 import InternalParcelNode from 'dataparcels/ParcelNode';
 import InternalAsNodes from 'dataparcels/arrange';
@@ -60,8 +60,8 @@ test('/ChangeRequest should export ChangeRequest', () => {
     expect(ChangeRequest).toBe(InternalChangeRequest);
 });
 
-test('/createUpdater should export createUpdater', () => {
-    expect(createUpdater).toBe(InternalCreateUpdater);
+test('/combine should export combine', () => {
+    expect(combine).toBe(InternalCreateUpdater);
 });
 
 test('/deleted should export deleted', () => {
