+++ packages/react-dataparcels/combine.js
@@ -1,0 +1,2 @@
+/* eslint-disable */
+module.exports = require('dataparcels/combine.js');
