--- packages/react-dataparcels/createUpdater.js
@@ -1,2 +1,0 @@
-/* eslint-disable */
-module.exports = require('dataparcels/createUpdater.js');
