+++ packages/react-dataparcels/package.json
@@ -19,7 +19,7 @@
     "Boundary.js",
     "cancel.js",
     "ChangeRequest.js",
-    "createUpdater.js",
+    "combine.js",
     "deleted.js",
     "ParcelBoundary.js",
     "ParcelDrag.js",
